File Manager Support for Folders  (#848)

+++ jarviscli/plugins/file_manager.py
@@ -7,7 +7,7 @@ from plugin import plugin
 @plugin("file manage")
 class file_manage:
     """"
-    Can manipulate files by deleting, moving, or renaming.
+    Can manipulate files and folders by deleting, moving, or renaming.
     """
 
     def __call__(self, jarvis, s):
@@ -20,6 +20,12 @@ class file_manage:
             self.move(jarvis, self.file)
         elif self.cmd == "rename":
             self.rename(jarvis, self.file)
+
+        # determine if directory entered is a file or folder
+        if os.path.isfile(self.file):
+            self.folder = False
+        else:
+            self.folder = True
 
     def get_file_directory(self, jarvis):
         self.file = jarvis.input("Enter the directory of the file you would like to edit: ")
@@ -50,33 +56,37 @@ class file_manage:
     def delete(self, jarvis, file):
         # function to delete files
 
-        # first, check if file exists
-        if os.path.exists(file):
+        if self.folder is False:
+            # first, check if file exists
+            if os.path.exists(file):
 
-            yes = True
-            while yes:
-                # confirm that file should be deleted
-                confirmation = jarvis.input("Are you sure you want to delete this file? This cannot be undone. (y/n)").lower()
+                yes = True
+                while yes:
+                    # confirm that file should be deleted
+                    confirmation = jarvis.input("Are you sure you want to delete this file? This cannot be undone. (y/n)").lower()
 
-                if confirmation == "y":
-                    try:
-                        # delete file
-                        os.remove(file)
-                    except:
-                        jarvis.say("Invalid file path")
+                    if confirmation == "y":
+                        try:
+                            # delete file
+                            if not self.folder:
+                                os.remove(file)
+                            else:
+                                os.rmdir(file)
+                        except:
+                            jarvis.say("Invalid file path")
 
-                    # break loop after removing file
-                    yes = False
+                        # break loop after removing file
+                        yes = False
 
-                elif confirmation == "n":
+                    elif confirmation == "n":
 
-                    # break loop if no confirmation
-                    yes = False
-                else:
-                    jarvis.say("Invalid input")
+                        # break loop if no confirmation
+                        yes = False
+                    else:
+                        jarvis.say("Invalid input")
 
-        else:
-            jarvis.say("file does not exist")
+            else:
+                jarvis.say("file does not exist")
 
     def move(self, jarvis, file):
         # function to move files
